backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
langchain==0.3.13
langchain-groq==0.2.3
langchain-core==0.3.28
httpx==0.28.1
